src/handlers/requirements.txt
pandas
pytz
numpy
boto3
sqlalchemy
sqlalchemy_utils
snowflake-sqlalchemy
